// FrontEnd/src/app/pages/Perfil/Model/editperfiladmin.ts
export interface userdtEdit {
    userName: string
    nombres: string
    apellidos: string
    fechaNacimiento: Date
}
export interface userdtperfil {
    userName: string
    nombres: string
    apellidos: string
    fechaNacimiento: Date
    fechaIngreso:Date
    mail:string
    identificacion:string
}




export interface admindtEdit {
    userName: string
    nombres: string
    apellidos: string
    fechaNacimiento: Date
}
export interface admindtperfil {
    userName: string
    nombres: string
    apellidos: string
    fechaNacimiento: Date
    fechaIngreso:Date
    mail:string
    identificacion:string
}